OvmfPkg: Move the 8259 PIC driver under 'CSM_ENABLE' build flag
The previous commit has updated the 8254TimerDxe to remove the dependency
on the Legacy8259 protocol and the 8259InterruptControllerDxe driver.

Also, we are going to view the Legacy8259 protocol and the 8259 driver as
CSM-specific components. And are expecting they will be moved to platforms
CSM folder (for those still have CSM support) before ultimately being
removed.

Thus, this commit will move the 8259 driver under the 'CSM_ENABLE' build
flag to mark it as CSM-specific.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Hao Wu <hao.a.wu@intel.com>