OvmfPkg/OvmfPkg.dec: Add PcdShellFile in OVMF DEC file
REF:https://bugzilla.tianocore.org/show_bug.cgi?id=1843

OVMF is currently using the PCD 'PcdShellFile' defined in
IntelFrameworkModulePkg.

This commit will add this PCD into the OVMF DEC file in order to drop
OVMF's dependency on IntelFrameworkModulePkg. The PCD token value 0x17 is
selected to fill the hole (seems the only one left) in OvmfPkg.

Please note that, instead of adding the PCD under section:
[PcdsFixedAtBuild, PcdsPatchableInModule]

as in IntelFrameworkModulePkg.dec file, it is added in section:
[PcdsFixedAtBuild]

in OvmfPkg.dec instead.

Cc: Ray Ni <ray.ni@intel.com>
Cc: Jordan Justen <jordan.l.justen@intel.com>
Cc: Laszlo Ersek <lersek@redhat.com>
Cc: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Signed-off-by: Hao A Wu <hao.a.wu@intel.com>